cleaned up duplicate files
