+++ index.android.js
@@ -1,9 +1,7 @@
 import React, { Component } from 'react';
 import { AppRegistry } from 'react-native';
 
-import IndexScene from './js/index.scene';
 import Setup from './js/setup';
-import SetupRoot from './js/setupRoot';
 
 
-AppRegistry.registerComponent('JustArrived', SetupRoot);
+AppRegistry.registerComponent('JustArrived', Setup);
