+++ index.ios.js
@@ -1,14 +1,7 @@
 import React, { Component } from 'react';
 import { AppRegistry } from 'react-native';
 
-import IndexScene from './js/index.scene';
+import Setup from './js/setup';
 
-class JustArrived extends Component {
-  render() {
-    return (
-      <IndexScene />
-    )
-  }
-}
 
-AppRegistry.registerComponent('JustArrived', () => JustArrived);
+AppRegistry.registerComponent('JustArrived', Setup);
