--- js/setupRoot.js
@@ -1,31 +1,0 @@
-
-import React, { Component } from 'react';
-import { Provider } from 'react-redux';
-
-import App from './app';
-import configureStore from './store/configureStore';
-
-function setup():React.Component {
-  class Root extends Component {
-
-    constructor() {
-      super();
-      this.state = {
-        isLoading: false,
-        store: configureStore(() => this.setState({ isLoading: false })),
-      };
-    }
-
-    render() {
-      return (
-        <Provider store={this.state.store}>
-          <App />
-        </Provider>
-      );
-    }
-  }
-
-  return Root;
-}
-
-export default setup;
